@@ -1,6 +1,6 @@
 {
   "name": "bfj",
-  "version": "7.0.2",
+  "version": "7.1.0",
   "description": "Big-friendly JSON. Asynchronous streaming functions for large JSON data sets.",
   "homepage": "https://gitlab.com/philbooth/bfj",
   "bugs": "https://gitlab.com/philbooth/bfj/issues",
@@ -29,18 +29,19 @@
     "node": ">= 8.0.0"
   },
   "dependencies": {
-    "bluebird": "^3.5.5",
-    "check-types": "^11.1.1",
+    "bluebird": "^3.7.2",
+    "check-types": "^11.2.3",
     "hoopy": "^0.1.4",
+    "jsonpath": "^1.1.1",
     "tryer": "^1.0.1"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
-    "eslint": "^6.0.1",
-    "mocha": "^6.1.4",
-    "please-release-me": "^2.1.2",
-    "proxyquire": "^2.1.0",
-    "request": "^2.88.0",
+    "axios": "^1.5.0",
+    "chai": "^4.3.8",
+    "eslint": "^8.48.0",
+    "mocha": "^10.2.0",
+    "please-release-me": "^2.1.4",
+    "proxyquire": "^2.1.3",
     "spooks": "^2.0.0"
   },
   "scripts": {
