@@ -1,6 +1,6 @@
 {
   "name": "jiti",
-  "version": "1.19.3",
+  "version": "1.20.0",
   "description": "Runtime typescript and ESM support for Node.js",
   "repository": "unjs/jiti",
   "license": "MIT",
@@ -21,60 +21,61 @@
     "lint": "eslint --ext .ts,.js . && prettier -c src lib test stubs",
     "lint:fix": "eslint --fix --ext .ts,.js . && prettier -w src lib test stubs",
     "release": "pnpm build && pnpm test && changelogen --release --push && npm publish",
-    "test": "pnpm lint && vitest run --coverage"
+    "test": "pnpm lint && vitest run --coverage && pnpm test:bun",
+    "test:bun": "bun --bun test test/bun"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.10",
-    "@babel/plugin-proposal-decorators": "^7.22.10",
-    "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.18.6",
-    "@babel/plugin-proposal-optional-chaining": "^7.21.0",
+    "@babel/core": "^7.22.15",
+    "@babel/plugin-proposal-decorators": "^7.22.15",
+    "@babel/plugin-transform-export-namespace-from": "^7.18.9",
+    "@babel/plugin-transform-nullish-coalescing-operator": "^7.18.6",
+    "@babel/plugin-transform-optional-chaining": "^7.21.0",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
     "@babel/plugin-syntax-import-assertions": "^7.22.5",
-    "@babel/plugin-transform-modules-commonjs": "^7.22.5",
-    "@babel/plugin-transform-typescript": "^7.22.10",
-    "@babel/preset-typescript": "^7.22.5",
-    "@babel/template": "^7.22.5",
-    "@babel/types": "^7.22.10",
+    "@babel/plugin-transform-modules-commonjs": "^7.22.15",
+    "@babel/plugin-transform-typescript": "^7.22.15",
+    "@babel/preset-typescript": "^7.22.15",
+    "@babel/template": "^7.22.15",
+    "@babel/types": "^7.22.15",
     "@types/babel__core": "^7.20.1",
     "@types/babel__template": "^7.4.1",
-    "@types/node": "^20.5.0",
-    "@types/object-hash": "^3.0.3",
+    "@types/node": "^20.5.9",
+    "@types/object-hash": "^3.0.4",
     "@types/resolve": "^1.20.2",
-    "@types/semver": "^7.5.0",
-    "@vitest/coverage-v8": "^0.34.2",
+    "@types/semver": "^7.5.1",
+    "@vitest/coverage-v8": "^0.34.3",
     "acorn": "^8.10.0",
     "babel-plugin-dynamic-import-node": "^2.3.3",
     "babel-plugin-parameter-decorator": "^1.0.16",
     "babel-plugin-transform-typescript-metadata": "^0.3.2",
-    "changelogen": "^0.5.4",
+    "changelogen": "^0.5.5",
     "config": "^3.3.9",
     "create-require": "^1.1.1",
     "destr": "^2.0.1",
     "escape-string-regexp": "^5.0.0",
-    "eslint": "^8.47.0",
+    "eslint": "^8.48.0",
     "eslint-config-unjs": "^0.2.1",
     "esm": "^3.2.25",
     "estree-walker": "^3.0.3",
-    "execa": "^7.2.0",
+    "execa": "^8.0.1",
     "fast-glob": "^3.3.1",
-    "mlly": "^1.4.0",
+    "mlly": "^1.4.2",
     "object-hash": "^3.0.0",
     "pathe": "^1.1.1",
     "pirates": "^4.0.6",
     "pkg-types": "^1.0.3",
-    "prettier": "^3.0.2",
+    "prettier": "^3.0.3",
     "reflect-metadata": "^0.1.13",
     "semver": "^7.5.4",
-    "std-env": "^3.3.3",
+    "std-env": "^3.4.3",
     "terser-webpack-plugin": "^5.3.9",
     "ts-loader": "^9.4.4",
-    "tslib": "^2.6.1",
-    "typescript": "^5.1.6",
+    "tslib": "^2.6.2",
+    "typescript": "^5.2.2",
     "vite": "^4.4.9",
-    "vitest": "^0.34.2",
+    "vitest": "^0.34.3",
     "webpack": "^5.88.2",
     "webpack-cli": "^5.1.4"
   },
-  "packageManager": "pnpm@8.6.12"
+  "packageManager": "pnpm@8.7.4"
 }
