@@ -1,9 +1,16 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import ReviewList from './ReviewList';
+import ReviewForm from './ReviewForm';
 import { getReviews } from '../api';
+
+const LIMIT = 6;
 
 function App() {
   const [order, setOrder] = useState('createdAt');
+  const [offset, setOffset] = useState(0);
+  const [hasNext, setHasNext] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
+  const [loadingError, setLoadingError] = useState(null);
   const [items, setItems] = useState([]);
   const sortedItems = items.sort((a, b) => b[order] - a[order]);
 
@@ -16,10 +23,40 @@ function App() {
     setItems(nextItems);
   };
 
-  const handleLoadClick = async () => {
-    const { reviews } = await getReviews();
-    setItems(reviews);
+  const handleLoad = async (options) => {
+    let result;
+    try {
+      setLoadingError(null);
+      setIsLoading(true);
+      result = await getReviews(options);
+    } catch (error) {
+      setLoadingError(error);
+      return;
+    } finally {
+      setIsLoading(false);
+    }
+
+    const { paging, reviews } = result;
+    if (options.offset === 0) {
+      setItems(reviews);
+    } else {
+      setItems((prevItems) => [...prevItems, ...reviews]);
+    }
+    setOffset(options.offset + options.limit);
+    setHasNext(paging.hasNext);
   };
+
+  const handleLoadMore = async () => {
+    await handleLoad({ order, offset, limit: LIMIT });
+  };
+
+  const handleSubmitSuccess = (review) => {
+    setItems((prevItems) => [review, ...prevItems]);
+  };
+
+  useEffect(() => {
+    handleLoad({ order, offset: 0, limit: LIMIT });
+  }, [order]);
 
   return (
     <div>
@@ -27,8 +64,14 @@ function App() {
         <button onClick={handleNewestClick}>최신순</button>
         <button onClick={handleBestClick}>베스트순</button>
       </div>
+      <ReviewForm onSubmitSuccess={handleSubmitSuccess} />
       <ReviewList items={sortedItems} onDelete={handleDelete} />
-      <button onClick={handleLoadClick}>불러오기</button>
+      {hasNext && (
+        <button disabled={isLoading} onClick={handleLoadMore}>
+          더 보기
+        </button>
+      )}
+      {loadingError?.message && <span>{loadingError.message}</span>}
     </div>
   );
 }
